api/app/pyproject.toml
[tool.poetry]
name = "Hackaton2023 Backend"
version = "0.0.1"
description = ""
authors = ["Kuzina Ekaterina"]

[tool.poetry.dependencies]
python = "^3.10"
Django = "^4.1.1"
djangorestframework = "^3.14.0"
django-filter = "^2.4.0"
django-cors-headers = "^3.7.0"
gunicorn = "^20.1.0"
SQLAlchemy = "^1.4.39"
psycopg2 = "^2.9.5"
pydantic = "^1.9.1"
bcrypt = "^4.0.1"
PyJWT = "^2.6.0"
pylint = "^2.16.0"
pika = "^1.3.1"
openpyxl = "^3.0.10"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
